requirements.txt
transformers
tqdm
pillow
streamlit
streamlit-shortcuts
pandas
scikit-learn
facenet-pytorch
ultralytics
faiss-cpu
ruff